feat: support using the action with Azure-hosted OpenAI models (#47)
This PR introduces a new, optional input to the GitHub Action,
`responses-api-endpoint`, which makes it possible to use this Action
with Azure. As explained in the updated `README.md`, if
`responses-api-endpoint` is specified, it will be the URL that Codex
hits to make Responses API requests. In practice, we expect this to be
used mostly by Azure users.

With the Azure account I just created, I got this working with the
following configuration:

```yaml
jobs:
  codex:
    runs-on: ubuntu-latest
    steps:
      - uses: actions/checkout@v5

      - name: Start Codex proxy
        uses: openai/codex-action@71690bb0c5f5196713197e9877f0b28950c7b415
        with:
          openai-api-key: ${{ secrets.AZURE_OPENAI_API_KEY }}
          responses-api-endpoint: "https://bolinfest-7804-resource.cognitiveservices.azure.com/openai/v1/responses"
```

Under the hood, if `responses-api-endpoint` is non-empty, it simply
becomes the `--upstream-url` argument to `codex-responses-api-proxy`.

TO DISCUSS: Better/shorter name than `responses-api-endpoint`?

This PR builds on the work of https://github.com/openai/codex/pull/6129.