Properly diff the PR against `master`
So when master moves forward, the new commits on master are also included in `base...head`, which makes it look like parts of the PR are “reverting” them.

When reviewing the PR, don't use the `base...head` notation (which includes all commits reachable from base or head BUT not from both). Instead, use `base..head` (commits reachable from head which are not reachable from base).